requirements.txt
fastapi
uvicorn[standard]
ultralytics
python-dotenv
python-multipart
